requirements.txt
chainlit
notebook
ipywidgets
tqdm
python-dotenv
openai
langchain